// frontend/src/api/models/scope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * The scope of a memory (i.e. how long it should be retained).
 *
 * LOCAL: Memory is only relevant to the current chapter.
 * RECENT: Memory is relevant to the current chapter and a few subsequent chapters.
 * PERSIST: Memory is relevant to all chapters and should be retained indefinitely until explicitly superseded or expired.
 */
export type Scope = typeof Scope[keyof typeof Scope];


export const Scope = {
  local: 'local',
  recent: 'recent',
  persist: 'persist',
} as const;